[DeadCode] Handle truthy && is_<type>() in RemoveRedundantTypeCheckRector
