+++ rules-tests/DeadCode/Rector/BinaryOp/RemoveRedundantTypeCheckRector/Fixture/array_param_truthy_check.php.inc
@@ -1,0 +1,31 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\BinaryOp\RemoveRedundantTypeCheckRector\Fixture;
+
+final class ArrayParamTruthyCheck
+{
+    public function buildWhereClauseFromArray($query, array $clauses, $expr = null): void
+    {
+        if ($clauses && is_array($clauses)) {
+            return;
+        }
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\BinaryOp\RemoveRedundantTypeCheckRector\Fixture;
+
+final class ArrayParamTruthyCheck
+{
+    public function buildWhereClauseFromArray($query, array $clauses, $expr = null): void
+    {
+        if ($clauses) {
+            return;
+        }
+    }
+}
+
+?>
