+++ rules-tests/DeadCode/Rector/BinaryOp/RemoveRedundantTypeCheckRector/Fixture/nullable_string_truthy_check.php.inc
@@ -1,0 +1,31 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\BinaryOp\RemoveRedundantTypeCheckRector\Fixture;
+
+final class NullableStringTruthyCheck
+{
+    public function run(?string $value): void
+    {
+        if ($value && is_string($value)) {
+            echo $value;
+        }
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\BinaryOp\RemoveRedundantTypeCheckRector\Fixture;
+
+final class NullableStringTruthyCheck
+{
+    public function run(?string $value): void
+    {
+        if ($value) {
+            echo $value;
+        }
+    }
+}
+
+?>
