+++ rules-tests/DeadCode/Rector/BinaryOp/RemoveRedundantTypeCheckRector/Fixture/skip_truthy_check_on_union.php.inc
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rector\Tests\DeadCode\Rector\BinaryOp\RemoveRedundantTypeCheckRector\Fixture;
+
+final class SkipTruthyCheckOnUnion
+{
+    public function run(array|string $value): void
+    {
+        if ($value && is_array($value)) {
+            return;
+        }
+    }
+}
