+++ rules-tests/DeadCode/Rector/BinaryOp/RemoveRedundantTypeCheckRector/config/configured_rule.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\Config\RectorConfig;
+use Rector\DeadCode\Rector\BinaryOp\RemoveRedundantTypeCheckRector;
+
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->rule(RemoveRedundantTypeCheckRector::class);
+};
