--- rules-tests/DeadCode/Rector/BooleanOr/RemoveRedundantNullableTypeCheckRector/config/configured_rule.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\DeadCode\Rector\BooleanOr\RemoveRedundantNullableTypeCheckRector;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(RemoveRedundantNullableTypeCheckRector::class);
-};
